pyproject.toml
[project]
name = "legislative-intelligence"
version = "0.1.0"
description = "A system for understanding the story of laws - linking statutes to their legislative history, regulatory implementation, and judicial interpretation"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    # Core
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",

    # Graph database
    "neo4j>=5.0.0",

    # Vector store & embeddings
    "pgvector>=0.2.0",
    "psycopg[binary]>=3.1.0",
    "openai>=1.0.0",  # For embeddings

    # Data processing
    "lxml>=5.0.0",
    "beautifulsoup4>=4.12.0",
    "pandas>=2.0.0",

    # Citation parsing
    "eyecite>=2.6.0",
    "regex>=2024.0.0",

    # API
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",

    # Utilities
    "rich>=13.0.0",
    "tqdm>=4.66.0",
    "tenacity>=8.2.0",

    # LLM / Anthropic
    "anthropic>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "ipython>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
